refactor Utils -> util
