+++ packages/salesforce/src/index.js
@@ -3,5 +3,6 @@ export { metadata };
 
 import * as Adaptor from './Adaptor';
 export default Adaptor;
-
 export * from './Adaptor';
+
+export * as util from './util';
